pyproject.toml
[tool.poetry]
name = "syto-api"
version = "0.1.0"
description = ""
authors = ["Przemek Kaliś <przemek.kalis@gmail.com>"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
isort = "^5.7.0"
black = "^20.8b1"
pre-commit = "^2.10.1"
pytest = "^5.2"
pytest-cov = "^2.11.1"
pytest-django = "^4.1.0"
python-dateutil = "^2.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
